Split api-related tests and fixtures
